Flag Initialization now performs an update. Should fix certain bugs
